requirements.txt
speechrecognition
pyaudio
sqlite3
ollama
smtplib
python-dotenv
